extracted reusable ListComponent shared by PostsList and MyPosts components
