+++ client/src/components/posts/postAuthor/PostAuthor.jsx
@@ -27,9 +27,6 @@ const PostAuthor = (props) => {
       <Avatar alt={props.author} src={props.authorAvatar} sx={styles.avatar} />
       <Typography variant="body2" component="p" sx={styles.author}>
         {"by "} {props.author ? props.author : "Unknown Author"} {""}
-        <Typography variant="body2" component="span" sx={{ fontWeight: 600 }}>
-          {`${props.quantity} ${props.quantity === 1 ? "post" : "posts"}`}
-        </Typography>
       </Typography>
     </Box>
   );
